fix:選択中のチンチラの表示をコンポーネント化
