+++ frontend/src/components/shared/Layout.jsx
@@ -1,4 +1,4 @@
-import { Header } from 'src/components/shared/Header'
+import { Header } from 'src/components/shared/Header/index'
 import { Footer } from 'src/components/shared/Footer'
 
 export const Layout = ({ children }) => {
